move video processing to task queue
